Modify: isolate attraction function
